refactor: Replace isValidLookback with asValidLookback
Returns the value itself (or undefined) instead of boolean, eliminating
the repeated guard pattern at every call site.
