+++ packages/jaeger-ui/src/utils/time-range-options.ts
@@ -95,9 +95,9 @@ export const TIME_RANGE_OPTIONS: readonly ITimeRangeOption[] = Object.freeze([
   },
 ]);
 
-/** Returns true when value is a known lookback string in TIME_RANGE_OPTIONS. */
-export function isValidLookback(value: string): boolean {
-  return TIME_RANGE_OPTIONS.some(o => o.lookback === value);
+/** Returns value when it is a known lookback string in TIME_RANGE_OPTIONS, otherwise undefined. */
+export function asValidLookback(value: string | undefined): string | undefined {
+  return TIME_RANGE_OPTIONS.some(o => o.lookback === value) ? value : undefined;
 }
 
 /**
